requirement.txt
langgraph
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
typing-extensions

sounddevice
scipy

python-Levenshtein

jupyter 
ipython 
langgraph-checkpoint 
langgraph-checkpoint-sqlite 
pygraphviz
pillow